fix(dump_regs): don't cast uint32_t regs to int8_t

+++ src/main.cc
@@ -26,7 +26,7 @@ dump_regs()
 {
   for (int i = 0; i < N_REGS / 4; ++i) {
     for (int j = 0; j < 4; ++j) {
-      printf("X%02d: %#-12X", i*4+j, static_cast<int8_t>(regfile[i*4+j]));
+      printf("X%02d: %#-12X", i*4+j, regfile[i*4+j]);
     }
     puts("");
   }
